D16_BoardApp/requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
bleach==4.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.3
django-allauth==0.61.1
django-apscheduler==0.6.2
django-filter==24.2
django-js-asset==2.2.0
django-summernote==0.8.20.0
idna==3.6
oauthlib==3.2.2
packaging==24.0
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzlocal==5.2
urllib3==2.2.1
webencodings==0.5.1
